requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.1
Markdown==3.7
numpy==2.1.1
pandas==2.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
referencing==0.35.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
